examples/crewai-qwen-demo/requirements.txt
crewai>=0.80.0
requests>=2.32.0
python-dotenv>=1.0.0
